refactor context engine
