+++ locales/zh-CN/home.json
@@ -1,6 +1,6 @@
 {
   "starter": {
-    "createAgent": "创建助理",
+    "createAgent": "创建 Agent",
     "deepResearch": "探究",
     "developing": "正在开发中",
     "image": "绘画",
