+++ packages/context-engine/src/index.ts
@@ -6,12 +6,12 @@ export { BaseProcessor } from './base/BaseProcessor';
 export { BaseProvider } from './base/BaseProvider';
 
 // Context Engine
+export * from './engine';
 export type { ContextEngineConfig } from './pipeline';
 export { ContextEngine } from './pipeline';
 
 // Context Providers
 export * from './providers';
-
 // Processors
 export type { PlaceholderValue, PlaceholderValueMap } from './processors';
 export {
@@ -27,19 +27,3 @@ export {
   ToolCallProcessor,
   ToolMessageReorder,
 } from './processors';
-
-// Tools Engine
-export type {
-  FunctionCallChecker,
-  GenerateToolsParams,
-  LobeToolManifest,
-  PluginEnableChecker,
-  ToolNameGenerator,
-  ToolsEngineOptions,
-  ToolsGenerationContext,
-  ToolsGenerationResult,
-} from './tools';
-export { filterValidManifests, ToolNameResolver, ToolsEngine, validateManifest } from './tools';
-
-// Messages Engine
-export * from './messages';
