+++ src/locales/default/home.ts
@@ -1,6 +1,6 @@
 export default {
   starter: {
-    createAgent: '创建助理',
+    createAgent: '创建 Agent',
     deepResearch: '探究',
     developing: '正在开发中',
     image: '绘画',
